Added files via upload
These are the notebooks I've been using, one of them is the cluster notebook (featured 13.2 results) and the other is local. I think I got scala pagerank working in it. I had to bring my cluster down though because my cost was spiralling up. In terms of testing the scala stuff, sbt needs to be installed (followed the instructions here: http://www.scala-sbt.org/release/docs/Setup.html) The results I got for the toy set are as follows:

(1,0.3040900819355781)
(2,2.837824310201201)
(3,2.4868526601543555)
(4,0.3625952805244703)
(5,0.7503400819355781)
(6,0.3625952805244703)
(7,0.15)
(8,0.15)
(9,0.15)
(10,0.15)
(11,0.15)

They seem to make sense but I'm not sure why some are greater than 1. It's the basic out of the box pagerank though...